initial commit (basic structure, does not compile)
